% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fav.R
\name{.fav_as_href}
\alias{.fav_as_href}
\title{Add data URI prefix to base64-encoded icon}
\usage{
.fav_as_href(fav_base64)
}
\arguments{
\item{fav_base64}{}
}
\value{
The base64-encoded icon with the data URI prefix.
}
\description{
Add data URI prefix to base64-encoded icon
}
\keyword{internal}
