% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/favawesome-package.R
\docType{package}
\name{favawesome-package}
\alias{favawesome}
\alias{favawesome-package}
\title{favawesome: What the Package Does}
\description{
The goal of favawesome is to...
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/shinyworks/favawesome}
  \item \url{https://shinyworks.github.io/favawesome/}
  \item Report bugs at \url{https://github.com/shinyworks/favawesome/issues}
}

}
\author{
\strong{Maintainer}: Jon Harmon \email{jonthegeek@gmail.com} (\href{https://orcid.org/0000-0003-4781-4346}{ORCID})

}
\keyword{internal}
