% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_favawesome_standalone.R
\name{use_favawesome_standalone}
\alias{use_favawesome_standalone}
\title{Add favawesome_icon Function to a Project}
\usage{
use_favawesome_standalone()
}
\value{
\code{NULL} invisibly (called for side effect of importing the
\code{favawesome_icon()} definition). \code{favawesome_icon()} returns the same value
as \code{\link[=fav]{fav()}}.
}
\description{
If you can't import \code{favawesome} in your project, you can still use a Font
Awesome (\url{https://fontawesome.com}) icon as the favicon (the icon that
appears on browser tabs). Calling this function will download the necessary
code to add a \code{favawesome_icon()} function to your project, and will ovewrite
the existing file if it already exists, making it easy to update the code.
The \code{favawesome_icon()} function takes the same arguments as \code{\link[=fav]{fav()}}.
}
