% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/favawesome-package.R
\docType{package}
\name{favawesome-package}
\alias{favawesome}
\alias{favawesome-package}
\title{favawesome: 'Font Awesome' Icons as 'Shiny' 'favicons'}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Easily use 'Font Awesome' icons as 'Shiny' 'favicons' (the icons that appear on browser tabs). 'Font Awesome' is a popular set of icons that can be used in web pages. 'favawesome' provides a simple way to use these icons as 'favicons' in 'Shiny' applications and other HTML pages.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://favawesome.shinyworks.org}
  \item \url{https://github.com/shinyworks/favawesome}
  \item Report bugs at \url{https://github.com/shinyworks/favawesome/issues}
}

}
\author{
\strong{Maintainer}: Jon Harmon \email{jonthegeek@gmail.com} (\href{https://orcid.org/0000-0003-4781-4346}{ORCID}) [copyright holder]

}
\keyword{internal}
