% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/favawesome-package.R
\docType{package}
\name{favawesome-package}
\alias{favawesome}
\alias{favawesome-package}
\title{favawesome: Use 'Font Awesome' Icons as 'Shiny' 'favicons'}
\description{
Easily and flexibly use 'Font Awesome' icons as 'Shiny' 'favicons'. 'Font Awesome' is a popular set of icons that can be used in web pages. 'favawesome' provides a simple way to use these icons as 'favicons' in 'Shiny' applications.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://shinyworks.github.io/favawesome/}
  \item \url{https://github.com/shinyworks/favawesome}
  \item Report bugs at \url{https://github.com/shinyworks/favawesome/issues}
}

}
\author{
\strong{Maintainer}: Jon Harmon \email{jonthegeek@gmail.com} (\href{https://orcid.org/0000-0003-4781-4346}{ORCID})

}
\keyword{internal}
